+++ sprout/csv/csv_validator.py
@@ -4,50 +4,17 @@ from typing import Mapping
 import polars as pl
 from polars import read_csv
 
-from sprout.csv_reader import _transform_to_suitable_csv_format
-
-
-class CsvValidationError:
-    """Information about a CSV validation error."""
-
-    def __init__(
-        self,
-        column: str,
-        data_type: pl.PolarsDataType,
-        row_number: int = 0,
-        value: str = "",
-        is_column_missing: bool = False,
-    ):
-        """CsvValidationError constructor.
-
-        Args:
-            column: Name of the column with the error
-            data_type: The expected data type of the column
-            row_number: The row of the error
-            value: The value that gave an error
-            is_column_missing: True if the entire column is missing
-        """
-        self.column = column
-        self.data_type = str(data_type.base_type())
-        self.row_number = row_number
-        self.value = value
-        self.is_column_missing = is_column_missing
-
-    def __str__(self):
-        """String representation of the CsvValidationError."""
-        if self.is_column_missing:
-            return f"Column '{self.column}' with type '{self.data_type}' is missing"
-
-        return (
-            f"Value '{self.value}' is not '{self.data_type}'. Column: "
-            f"'{self.column}', row: '{self.row_number}'"
-        )
+from sprout.csv.csv_reader import _transform_to_suitable_csv_format
+from sprout.csv.csv_validation_error import CsvValidationError
 
 
 def validate_csv(
     file_path: str, type_mapping: Mapping[str, pl.PolarsDataType]
 ) -> list[CsvValidationError]:
-    """Validate the CSV file.
+    """Validates the CSV file.
+
+    Validates the CSV file based on the type_mapping. Every validation error creates
+    a CsvValidationError object.
 
     Args:
         file_path: The path of the CSV file to validate.
