+++ sprout/tests/tests_csv_validator.py
@@ -5,7 +5,7 @@ from unittest import TestCase
 
 import polars as pl
 
-from sprout.csv_validator import validate_csv
+from sprout.csv.csv_validator import validate_csv
 
 
 class CsvValidatorTests(TestCase):
